Remove SSZ encoding tests. Ensure all test classes are in their own file, not as internal classes.
