+++ bls/src/test/java/tech/pegasys/teku/bls/MikuliSignatureTest.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2020 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.bls;
+
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
+import tech.pegasys.teku.bls.impl.mikuli.MikuliBLS12381;
+
+public class MikuliSignatureTest extends BLSSignatureTest {
+
+  @BeforeAll
+  public static void init() {
+    BLS.setBlsImplementation(MikuliBLS12381.INSTANCE);
+  }
+
+  @AfterAll
+  public static void cleanup() {
+    BLS.resetBlsImplementation();
+  }
+}
